+++ kurtosis-module/impl/participant_network/cl/lighthouse/lighthouse_launcher.go
@@ -21,8 +21,8 @@ const (
 
 	// ---------------------------------- Beacon client -------------------------------------
 	consensusDataDirpathOnBeaconServiceContainer = "/consensus-data"
-	beaconConfigDataDirpathRelToSharedDirRoot    = "config-data"
-	sharedJWTSecretRelFilepath                   = "jwtsecret"
+	beaconConfigDataDirpathRelToSharedDirRoot = "config-data"
+	jwtSecretFilepathRelToSharedDirRoot       = "jwtsecret"
 
 	// Port IDs
 	beaconTcpDiscoveryPortID = "tcpDiscovery"
@@ -76,10 +76,11 @@ var lighthouseLogLevels = map[module_io.GlobalClientLogLevel]string{
 type LighthouseCLClientLauncher struct {
 	// The dirpath on the module container where the CL genesis config data directory exists
 	configDataDirpathOnModuleContainer string
+	jwtSecretFilepathOnModuleContainer string
 }
 
-func NewLighthouseCLClientLauncher(configDataDirpathOnModuleContainer string) *LighthouseCLClientLauncher {
-	return &LighthouseCLClientLauncher{configDataDirpathOnModuleContainer: configDataDirpathOnModuleContainer}
+func NewLighthouseCLClientLauncher(configDataDirpathOnModuleContainer string, jwtSecretFilepathOnModuleContainer string) *LighthouseCLClientLauncher {
+	return &LighthouseCLClientLauncher{configDataDirpathOnModuleContainer: configDataDirpathOnModuleContainer, jwtSecretFilepathOnModuleContainer: jwtSecretFilepathOnModuleContainer}
 }
 
 func (launcher *LighthouseCLClientLauncher) Launch(
@@ -199,9 +200,9 @@ func (launcher *LighthouseCLClientLauncher) getBeaconContainerConfigSupplier(
 			)
 		}
 
-		JWTSecretRelFilepath := sharedDir.GetChildPath(sharedJWTSecretRelFilepath)
-		if err := service_launch_utils.CopyFileToSharedPath(launcher.configDataDirpathOnModuleContainer, JWTSecretRelFilepath); err != nil {
-			return nil, stacktrace.Propagate(err, "An error occurred copying JWT secret file '%v' into shared directory path '%v'", launcher.configDataDirpathOnModuleContainer, sharedJWTSecretRelFilepath)
+		jwtSecretSharedPath := sharedDir.GetChildPath(jwtSecretFilepathRelToSharedDirRoot)
+		if err := service_launch_utils.CopyFileToSharedPath(launcher.jwtSecretFilepathOnModuleContainer, jwtSecretSharedPath); err != nil {
+			return nil, stacktrace.Propagate(err, "An error occurred copying JWT secret file '%v' into shared directory path '%v'", launcher.jwtSecretFilepathOnModuleContainer, jwtSecretFilepathRelToSharedDirRoot)
 		}
 
 		elClientRpcUrlStr := fmt.Sprintf(
@@ -210,10 +211,10 @@ func (launcher *LighthouseCLClientLauncher) getBeaconContainerConfigSupplier(
 			elClientCtx.GetRPCPortNum(),
 		)
 
-		executionClientRpcUrlStr := fmt.Sprintf(
+		elClientEngineRpcUrlStr := fmt.Sprintf(
 			"http://%v:%v",
 			elClientCtx.GetIPAddress(),
-			elClientCtx.GetEnginePortNum(),
+			elClientCtx.GetEngineRPCPortNum(),
 		)
 
 		configDataDirpathOnService := configDataDirpathOnServiceSharedPath.GetAbsPathOnServiceContainer()
@@ -249,9 +250,9 @@ func (launcher *LighthouseCLClientLauncher) getBeaconContainerConfigSupplier(
 			//   https://github.com/sigp/lighthouse/blob/7c88f582d955537f7ffff9b2c879dcf5bf80ce13/scripts/local_testnet/beacon_node.sh
 			// and the option says it's "useful for testing in smaller networks" (unclear what happens in larger networks)
 			"--disable-packet-filter",
-			"--execution-endpoints=" + executionClientRpcUrlStr,
+			"--execution-endpoints=" + elClientEngineRpcUrlStr,
 			"--eth1-endpoints=" + elClientRpcUrlStr,
-			"--jwt-secrets=" + JWTSecretRelFilepath.GetAbsPathOnServiceContainer(),
+			"--jwt-secrets=" + jwtSecretSharedPath.GetAbsPathOnServiceContainer(),
 			// Set per Paris' recommendation to reduce noise in the logs
 			"--subscribe-all-subnets",
 			// vvvvvvvvvvvvvvvvvvv METRICS CONFIG vvvvvvvvvvvvvvvvvvvvv